Create CODE_OF_CONDUCT.md
I got blasted for being too harsh last year so I think it's important to have this as a reminder to all of us to be civil. I just want us to be smart, and I will point out when we aren't smart. I will criticize mistakes and failures, and it might sound mean, but don't take it personally, we are here to program a robot, we can't let our feelings get in the way. We all need to be able to take some heat. That includes the mentors who are overseeing this project. I will be straightforward and tough on all of you, and I am prepared for the same from all of you.